task_manager/requirement.txt
Django==5.0.4
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
python-dotenv==1.1.1